feat: Enhance project management features and update data models
- Added new endpoints in ProjectController for creating and updating projects.
- Introduced ProjectRequestDTO for project creation and updates.
- Implemented getFeaturedProjects method in ProjectService to retrieve featured projects.
- Updated ProjectMapper to handle DTO to entity conversions and vice versa.
- Changed ID types from Integer to Long in various DTOs and entity classes for consistency.
- Updated database schema generation strategy to 'create' in application.yaml for fresh setups.
- Refactored existing entities (About, Education, Experience, FAQ, Hobby, PersonalInfo, Project, Skill, Testimonial, Value) to use Long as ID type.
- Adjusted repository interfaces to reflect the new ID type.

+++ backend/src/main/java/br/com/wallace/portfolio/model/entities/About.java
@@ -1,10 +1,21 @@
 package br.com.wallace.portfolio.model.entities;
 
-import jakarta.persistence.*;
+import java.util.List;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.CollectionTable;
+import jakarta.persistence.Column;
+import jakarta.persistence.ElementCollection;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import java.util.List;
 
 @Entity
 @Table(name = "about")
@@ -14,7 +25,7 @@ import java.util.List;
 public class About {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(columnDefinition = "TEXT")
     private String introduction;
