+++ backend/src/main/java/br/com/wallace/portfolio/model/repository/AboutRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.wallace.portfolio.model.entities.About;
 
-public interface AboutRepository extends JpaRepository<About, Integer> {
+public interface AboutRepository extends JpaRepository<About, Long> {
     
 }
